+++ tests/Configuration.cpp
@@ -12,6 +12,7 @@
 #include <MicroOcpp/Core/Context.h>
 #include <MicroOcpp/Operations/CustomOperation.h>
 #include <MicroOcpp/Core/SimpleRequestFactory.h>
+#include <MicroOcpp/Debug.h>
 
 using namespace MicroOcpp;